Base-setup/requirements.txt
aniso8601==7.0.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-import-export==1.2.0
django-taggit==1.1.0
et-xmlfile==1.0.1
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.1
promise==2.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3

Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
youtube-dl==2019.11.5
